Added a Lockscreen Text Color Summary
Patchset 2 Changed the summary to be more concise. This is a real head
scratcher so if anyone has any hints as to how to best describe this
please let me know...

As lockscreen text color modifies more than just LS text now...it also
changes the LS weather & Calendar...If you guys think of a better way of phrasing this please
comment or submit a new patchset. Also removed some whitespace I found.

Change-Id: I9f03d4ad7fa1027bffa2d900625d2379b2c33279